requirements.txt
pystray>=0.19.4
Pillow>=10.1.0